==> 1_Data_download.ipynb <==
# Download data
### Load packages

import yfinance as yf
import pandas as pd
import re

### Define timeframe

start = '2010-05-28'
end = '2023-03-21'

### Download data
#### DAX index

# Download index data
index_data_dax = yf.download('^GDAXI', start, end)

#### DAX constituents

# Read constituents table from wikipedia
#wiki_dax = pd.read_html('https://en.wikipedia.org/wiki/DAX')
#constituents_table_dax = wiki_dax[4]
#constituents_symbols_dax = constituents_table_dax['Ticker'].to_list()

# For reproducing
constituents_symbols_dax = ['ADS.DE', 'AIR.DE', 'ALV.DE', 'BAS.DE', 'BAYN.DE', 'BEI.DE', 'BMW.DE', 'BNR.DE', 'CBK.DE', 'CON.DE', '1COV.DE', 'DTG.DE', 'DBK.DE', 'DB1.DE', 'DPW.DE', 'DTE.DE', 'EOAN.DE', 'FRE.DE', 'HNR1.DE', 'HEI.DE', 'HEN3.DE', 'IFX.DE', 'MBG.DE', 'MRK.DE', 'MTX.DE', 'MUV2.DE', 'P911.DE', 'PAH3.DE', 'QIA.DE', 'RHM.DE', 'RWE.DE', 'SAP.DE', 'SRT3.DE', 'SIE.DE', 'ENR.DE', 'SHL.DE', 'SY1.DE', 'VOW3.DE', 'VNA.DE', 'ZAL.DE']

# Download constituents data
constituents_data = yf.download(constituents_symbols_dax, start, end)

# Remove constituents with missing data
constituents_data = constituents_data.dropna(axis = 1)

# Remove ".DE" from Symbols
constituents_data.rename(columns = lambda x: re.sub('.DE', '', x.strip()), inplace = True)

# Drop wrong dates by index data
constituents_data = constituents_data.drop(list(sorted(set(constituents_data.index.to_list()) - set(index_data_dax.index.to_list()))), axis = 'index')

#### Enhanced timeframe for technical indicators

# Set new start
start = '2010-03-27'

# Same
constituents_data_ta = yf.download(constituents_symbols_dax, start, end)
index_data_dax_ta = yf.download(dax, start, end)
constituents_data_ta = constituents_data_ta.dropna(axis = 1)
constituents_data_ta.rename(columns = lambda x: re.sub('.DE', '', x.strip()), inplace = True)
constituents_data_ta = constituents_data_ta.drop(list(sorted(set(constituents_data_ta.index.to_list()) - set(index_data_dax_ta.index.to_list()))), axis = 'index')

### Save data

# DAX constituents
constituents_data.to_csv('data/constituents_data.csv')

# DAX constituents for technical indicators
constituents_data_ta.to_csv('data/constituents_data_ta.csv')
